+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [Unreleased]
+- Fix peripheral inheriting size, accessType and resetValue from device
+
 ## [v1.6.0] - 2025-08-30
 - PR[#11](https://github.com/mcu-debug/peripheral-viewer/issues/11): An alternative way to load peripherals information (see [extending-peripheral-viewer.md](https://github.com/mcu-debug/peripheral-viewer/blob/main/docs/extending-peripheral-viewer.md))